8151413: os::allocation_granularity/page_size and friends return signed values

+++ src/hotspot/cpu/ppc/gc/z/zGlobals_ppc.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, 2023, Oracle and/or its affiliates. All rights reserved.
  * Copyright (c) 2021 SAP SE. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
@@ -107,7 +107,7 @@ static unsigned int probe_valid_max_address_bit(size_t init_bit, size_t min_bit)
   unsigned int max_valid_address_bit = 0;
   void* last_allocatable_address = nullptr;
 
-  const unsigned int page_size = os::vm_page_size();
+  const size_t page_size = os::vm_page_size();
 
   for (size_t i = init_bit; i >= min_bit; --i) {
     void* base_addr = (void*) (((unsigned long) 1U) << i);
